# Bob performance optimizations for sequential builds