--- a/com.dynamo.cr/com.dynamo.cr.bob.test/src/com/dynamo/bob/pipeline/ExtenderUtilTest.java
+++ b/com.dynamo.cr/com.dynamo.cr.bob.test/src/com/dynamo/bob/pipeline/ExtenderUtilTest.java
@@ -77,7 +77,7 @@ public class ExtenderUtilTest {
 
         project = new Project(fileSystem, tmpDir.getAbsolutePath(), "build/default");
 
-        project.loadProjectFile();
+        project.loadProjectFile(true);
     }
 
     @After