--- a/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Bob.java
+++ b/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Bob.java
@@ -643,6 +643,7 @@ public class Bob {
     }
 
     private static void setupProject(Project project, boolean resolveLibraries, IProgress progress) throws IOException, LibraryException, CompileExceptionError {
+        project.loadProjectFile(false);
         BobProjectProperties projectProperties = project.getProjectProperties();
         String[] dependencies = projectProperties.getStringArrayValue("project", "dependencies");
         List<URL> libUrls = new ArrayList<>();
@@ -820,10 +821,6 @@ public class Bob {
                     project.addBuildServerHeader(header);
                 }
             }
-            TimeProfiler.stop();
-
-            TimeProfiler.start("loadProjectFile");
-            project.loadProjectFile();
             TimeProfiler.stop();
 
             TimeProfiler.start("setupProject");