--- a/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
+++ b/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
@@ -2262,7 +2262,7 @@ public class Project {
         TimeProfiler.start("findResourcePathsByExtension");
         TimeProfiler.addData("path", _path);
         TimeProfiler.addData("ext", ext);
-
+        _path = ".".equals(_path) ? "" : _path;
         final String path = Project.stripLeadingSlash(_path);
 
         // Initialize and cache all paths only once