--- a/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
+++ b/com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/Project.java
@@ -192,6 +192,11 @@ public class Project {
         clearProjectProperties();
     }
 
+    // For tests
+    public void cleanupResourcePathsCache() {
+        allResourcePathsCache = null;
+    }
+
     public void dispose() {
         this.fileSystem.close();
     }