This fix speeds up Bob.jar initialization, which is especially useful for sequential builds without many changes (e.g. Cmd+M to Build HTML5 from the editor, or bundling).  
* About 4 times faster initialization. In the test setup, it's 400 ms vs 1600 ms (for a project with a large file count, it's 3.5 s vs 20 s)  
* Speed-up of the task creation stage for a small project in the test setup: 900 ms vs 2000 ms (for a large project, about 5%)

Fix https://github.com/defold/defold/issues/10854

# Technical details


## `findResourcePaths`
Reading through the code, `findResourcePaths` appears suspicious.  
A dedicated scope in the build time profiler, highlighted in green, shows the usage of this function and the time it takes.

<img width="2878" height="1436" alt="CleanShot 2025-07-15 at 14 34 52@2x" src="https://github.com/user-attachments/assets/6d6c4bc8-7e61-459c-9298-4b5dbc0415bb" />
Profiler shows that `findResourcePaths` is called multiple times, each time scanning the file system.  
The `findResourcePaths()` method internally calls `findResourcePathsByExtension()`, where a file extension can be specified.  
In this test project, that function is called 7+ times, resulting in at least 20 full scan operations during a build (depending on project setup).

Given that `bob.jar` doesn’t modify the project folder during the build, it makes sense to scan the disk once and reuse the results.  
A cache was implemented at the `findResourcePathsByExtension()` level to avoid redundant scans.

**Result**:  
This change significantly speeds up the build process — about 5x faster for the test project.

<img width="2896" height="1866" alt="CleanShot 2025-07-15 at 17 09 06@2x" src="https://github.com/user-attachments/assets/6e422ddb-7161-4cbd-8768-4dcac372678e" />

## doScan
Further investigation shows that our Java class scanner calls all the static initializers, which isn't necessary at this stage.  
Excluding such classes from the scan stage helped reduce `doScan` time by 2x — from about 1s to 300–500 ms.  
Excluding the `com.dynamo.bob.pipeline.antlr` namespace further reduced scan time to about 160–230 ms.

## Ogg Validation
The Defold build pipeline has two main stages: task creation and task building.  
Profiling showed anomalously long creation times for `.ogg` files. Investigation revealed that `.ogg` validation occurred during the task creation stage, which is unnecessary.  
Moving validation to the task building stage saved about 500 ms during creation — important for sequential builds.

## Various Optimizations
- By default, we have a fairly large gamepads file that doesn't reference other resources.  
  Calling `createSubTasks` on it was a waste of time. Overriding the default `create` method saved about 120 ms on each iterative build.
- Lua parser is one of the heaviest operations during task creation.  
  Small improvements in the parser reduced parsing time by about 10%.

**Result**:  
All these improvements helped reduce sequential build time on a test project from **3.6s to 1.4s** (on average).