@AGulev 2025-07-16
One more comparison from a real project on another machine (sequential build):
5 seconds before (you may see that it takes a lot of time to scan disk to detect that native extensions don't need to be rebuilt)
<img width="2468" height="1220" alt="CleanShot 2025-07-16 at 22 41 21@2x" src="https://github.com/user-attachments/assets/ab24bacc-1fc4-449f-bf65-3b44a38f88a7" />
1 second one
<img width="2488" height="1518" alt="CleanShot 2025-07-16 at 22 42 13@2x" src="https://github.com/user-attachments/assets/dd3825c7-9703-4c62-a050-51eb23bec1a6" />